fix stacked envs, address comments
